Refactor: Remove EditCategoryModal and TerminateInstructor components; add new CoursesContent, CourseFilterBar, CourseFilterItem, CoursesTab, ReviewCard, AdminStats, BecomeInstructorInfoModal, and updated EditCategoryModal and TerminateInstructor components with improved functionality and structure.
