+++ components/course/course-content.jsx
