+++ components/dashboard/admin/terminate-instructor.jsx
